requirements.txt
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
idna==2.8
joblib==0.13.2
kiwisolver==1.1.0
matplotlib==3.1.0
numpy==1.16.3
pandas==0.24.2
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
requests==2.22.0
scikit-learn==0.21.1
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
sklearn==0.0
urllib3==1.25.2
